[src-server] Higher rated players may not be allowed to decline challenges
And force the time control to be chosen at the moment of sending the
challenge.

Added a configurable option to allow or disallow higher rated players to
decline or not a challenge sent to them by a lower rated player.

That is, when player A challenges player B to a game of a specific time
control, B may not decline the challenge if B is higher rated than A.
This behavior is configurable bia the configuration.json file.